Move filtering of returned resources to certnames
Currently, the resources returned from a resources
query are munged into a list of the hostnames that
returned those resources.

This makes it hard to re-use the query_puppetdb
method for cases where you actually want all
of the resources returned by the resources rest call.

This commit ensures that query_puppetdb always just
returns what the rest calls return and moves the
munging to certnames into the find_nodes_matching
method.