add allow conflicts parameter
I had previously added code that raised an exception
on parameter conflicts. This is good for the general
case, but having resources that are not exactly the
same everywhere is totally valid.

It makes sense to allow users to be able to
query for duplicate resources.

I am a little concerned about  the fact that this
option changes the format of what is returned,
but I'll sort that out later.