refactor query_resource function
This refactor allows a user to specify a filter
without having to specify a query.

This was done so that users did not have to specify
the resource reference twice (once for the query,
once for the filter)

This implementation works, but it calls the DB way
too many times.