Refactor functions to not use face as API
After some pain, I realized that trying to call
the face from the functions that I wrote was a design
mistake.

This is because I was having troubling getting the
faces to load from a relative path.

This massive refactor resolves the issue so that the 
puppet functions just work (that everything can just
be loaded from the module)

- All logic is removed from the faces. The faces
are just the CLI wrapper to the puppetdb query
methods
- The functions are changed to use the puppetdb
methods instead of the face.
- All requires that are internal to the module have
been changed to be relative path lookups.