Support querying for nested facts
Facter 3 introduced a lot of facts that are large hashes, which gets
difficult to work with in Puppet when using query_nodes or query_facts.
The types don't seem to coerce just right, so you'll get back a hash of
values with the type `Runtime[ruby string]`, which Puppet doesn't know
what to do with at the manifest level.

This PR introduces the ability to pass an array of keys to query out of
a fact hash. You can do something like:
```puppet
  query_nodes('<node_query>', ['networking', 'interfaces', 'eth0', 'network'])
```
to get back eth0 network for all matching nodes.

It's implemented for both query_nodes and query_facts in this PR.