gemspec: Remove puppet dependency
I feel horrible for doing this but with the AIO packaging this causes problems.

The Puppet gem has a runtime dependency on Facter, which means it'll try and pull in Facter. However, if you try to install this gem in the Puppet or Puppet Enterprise Ruby environment you'll be greeted with the wonderful error message: `facter" from facter conflicts with /opt/puppetlabs/puppet/bin/facter`.

It's for this same reason that puppetlabs_spec_helper and a few others have stopped explicitly depending on Puppet. It's terrible as installing it in a pure environment would result in a not working package.